requirements.txt
python-socketio==5.6.0
sanic==22.3.1
sanic-cors==2.0.1
loguru==0.6.0